% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{any_NA_dup_outliers}
\alias{any_NA_dup_outliers}
\title{any_NA_dup_outliers}
\usage{
any_NA_dup_outliers(df, tabular = FALSE)
}
\arguments{
\item{df}{data.frame object}

\item{tabular}{= FALSE}
}
\value{
tribble of truths, or kable-styled table if tabular=TRUE
}
\description{
Inspects the given data.frame for NAs, duplicates, outliers, and extreme values.
If the data.frame is not fully numeric, this attempts to make it so in order to
run 'rstatix::is_outlier' and 'is_extreme'.
}
