% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovX.R
\name{aov_plot}
\alias{aov_plot}
\title{aov_plot}
\usage{
aov_plot(df, formula)
}
\arguments{
\item{df}{data.frame object}

\item{formula}{valid formula object}
}
\description{
Given a data.frame and valid formula, this plots Mean.Sq vs F.value
from an Analysis of Variance (aov) summary.  If the data.frame is not
fully numeric, this attempts to make it so in order to run 'aov'.
}
