% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_numeric_df}
\alias{as_numeric_df}
\title{as_numeric_df}
\usage{
as_numeric_df(df)
}
\arguments{
\item{df}{data.frame object}
}
\value{
df mutated
}
\description{
For a given data.frame, this mutates each non-numeric column to instead contain
the indices of that column's vectorized unique values, which are then scaled without
centering, to minimize variance while remaining positive.
}
