% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrX.R
\name{corr_plot}
\alias{corr_plot}
\title{corr_plot}
\usage{
corr_plot(df, digits = 2, sig.level = 0.001)
}
\arguments{
\item{df}{data.frame object}

\item{digits}{= 2}

\item{sig.level}{= 0.001}
}
\description{
Wrapper around ggcorrplot, to plot the correlations for the given data.frame.
If the provided data.frame is not fully numeric, this attempts to make it so
in order to run 'cor'.
}
