% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.X-package.R
\docType{package}
\name{data.frame.X-package}
\alias{data.frame.X}
\alias{data.frame.X-package}
\title{data.frame.X: Simplified visualization methods for a 'data.frame'}
\description{
Intended to be an ever-growing collection of methods pertaining to analyzing and plotting a 'data.frame'.
}
\details{
data.frame.X provides custom functions to simplify common analysis and visualization
tasks against a data.frame.  Many analysis tools require a fully numeric data.frame,
and so this provides \code{as_numeric_df}.  And for quick, initial looks at Principal
Component Analysis, correlation, and variance, this provides \code{pca_plot}, \code{corr_plot},
\code{aov_plot} and \code{nzv_plot}.
}
\references{
A portion of \code{pca_plot} was derived from this \href{https://www.sthda.com/english/articles/31-principal-component-methods-in-r-practical-guide/118-principal-component-analysis-in-r-prcomp-vs-princomp}{Statistical Tools for High-Throughput Data Analysis} article.
}
\author{
\strong{Maintainer}: hbmaclean \email{hbmacl@gmail.com}

}
