% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompX.R
\name{.pca_barplot}
\alias{.pca_barplot}
\title{.pca_barplot}
\usage{
.pca_barplot(prcomp_obj, PCs)
}
\arguments{
\item{prcomp_obj}{prcomp object}

\item{PCs}{indices of PCs that explain the wanted portion of variance}
}
\description{
Private method to render a barplot from the provided prcomp object.
}
