% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompX.R
\name{.pca_boxplot}
\alias{.pca_boxplot}
\title{.pca_boxplot}
\usage{
.pca_boxplot(prcomp_obj, PCs)
}
\arguments{
\item{prcomp_obj}{prcomp object}

\item{PCs}{indices of PCs that explain the wanted portion of variance}
}
\description{
Private method to render a boxplot from the provided prcomp object.
}
\seealso{
\url{https://www.sthda.com/english/articles/31-principal-component-methods-in-r-practical-guide/118-principal-component-analysis-in-r-prcomp-vs-princomp}
}
