% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompX.R
\name{.pca_screeplot}
\alias{.pca_screeplot}
\title{.pca_screeplot}
\usage{
.pca_screeplot(prcomp_obj, PCs)
}
\arguments{
\item{prcomp_obj}{prcomp object}

\item{PCs}{indices of PCs that explain the wanted portion of variance}
}
\description{
Private method to render a scree plot from the provided prcomp object.
}
\seealso{
\url{https://www.rdocumentation.org/packages/factoextra/versions/1.0.7/topics/eigenvalue}
}
