% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{install_packages}
\alias{install_packages}
\title{install_packages}
\usage{
install_packages(packages)
}
\arguments{
\item{packages}{vector of packages}
}
\value{
implicit result of installations
}
\description{
Wrapper around the common !require/install/library pattern,
to install and load a given vector of packages.
}
\details{
Stops with error if packages are not provided as a character vector.
}
