% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_numeric_df}
\alias{is_numeric_df}
\title{is_numeric_df}
\usage{
is_numeric_df(df)
}
\arguments{
\item{df}{data.frame object}
}
\value{
TRUE/FALSE whether df is a fully numeric data.frame
}
\description{
Checks whether the given df is indeed a data.frame,
and that all columns are numeric.
}
