% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearZeroVarX.R
\name{nzv_plot}
\alias{nzv_plot}
\title{nzv_plot}
\usage{
nzv_plot(df)
}
\arguments{
\item{df}{data.frame object}

\item{tabular}{= FALSE}
}
\description{
This calls nearZeroVar, and blends the results with a variability calculated
as number of distinct values per column divided by row count.  If the data.frame
is not fully numeric, this attempts to make it so in order to run 'nearZeroVar'.
}
