% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompX.R
\name{pca_plot}
\alias{pca_plot}
\title{pca_plot}
\usage{
pca_plot(df, pct = 0.9, which = c("bar", "scree", "box"))
}
\arguments{
\item{df}{data.frame object}

\item{pct}{= 0.90}

\item{which}{= c('bar', 'scree', 'box')}
}
\description{
Arranges up to three plots for princicpal components explaining the wanted
percentage of variance for the given data.frame.  If the provided data.frame
is not fully numeric, this attempts to make it so in order to run 'prcomp'.
}
\details{
The process stops if zero PCs are identified, presumably due to too low a
value for pct.
}
\seealso{
\url{https://www.rdocumentation.org/packages/factoextra/versions/1.0.7/topics/eigenvalue}

\url{https://www.sthda.com/english/articles/31-principal-component-methods-in-r-practical-guide/118-principal-component-analysis-in-r-prcomp-vs-princomp}
}
